@@ -2,57 +2,71 @@
   "name": "bonik",
   "version": "3.0.0",
   "private": true,
+  "type": "module",
+  "license": "https://themeforest.net/licenses/standard",
   "scripts": {
-    "dev": "next dev --port 9000",
     "build": "next build",
+    "db:push": "drizzle-kit push",
+    "db:studio": "drizzle-kit studio",
+    "db:introspect": "drizzle-kit introspect",
+    "dev": "next dev --port 9000",
+    "lint": "next lint",
     "start": "next start",
-    "lint": "next lint"
+    "supabase:gen": "bunx supabase gen types typescript --project-id 'gddczfseozyqgzfnkeft' --schema public > src/lib/supabase/types.ts"
   },
   "dependencies": {
     "@emotion/is-prop-valid": "^1.2.2",
     "@radix-ui/react-slot": "^1.0.2",
     "@styled-system/theme-get": "^5.1.2",
     "@supabase/ssr": "latest",
-    "axios": "^1.6.7",
+    "@t3-oss/env-nextjs": "^0.10.1",
+    "@trpc/client": "next",
+    "@trpc/react-query": "next",
+    "@trpc/server": "next",
+    "axios": "^1.6.8",
     "axios-mock-adapter": "^1.22.0",
-    "chart.js": "^4.4.1",
+    "chart.js": "^4.4.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
-    "date-fns": "^3.3.1",
+    "date-fns": "^3.6.0",
+    "drizzle-orm": "^0.30.10",
     "embla-carousel-react": "^8.0.4",
-    "formik": "^2.4.5",
+    "formik": "^2.4.6",
     "lodash": "^4.17.21",
     "lucide-react": "^0.378.0",
-    "next": "14.1.0",
+    "next": "14.2.3",
     "nprogress": "^0.2.0",
+    "postgres": "^3.4.4",
     "pure-react-carousel": "^1.30.1",
-    "react": "18.2.0",
+    "react": "18.3.1",
     "react-chartjs-2": "^5.2.0",
     "react-click-away-listener": "^2.2.3",
-    "react-dom": "18.2.0",
+    "react-dom": "18.3.1",
     "react-dropzone": "^14.2.3",
     "react-html-props": "^2.0.9",
-    "react-icons": "^5.0.1",
+    "react-icons": "^5.2.1",
     "react-modal": "^3.16.1",
     "react-paginate": "^8.2.0",
     "react-scroll": "^1.9.0",
     "react-select": "^5.8.0",
-    "react-svg": "^16.1.33",
-    "sharp": "^0.33.2",
-    "simplebar-react": "^3.2.4",
-    "styled-components": "^6.1.8",
+    "react-svg": "^16.1.34",
+    "sharp": "^0.33.4",
+    "simplebar-react": "^3.2.5",
+    "styled-components": "^6.1.11",
     "styled-system": "^5.1.5",
+    "superjson": "^2.2.1",
     "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
-    "yup": "^1.3.3"
+    "yup": "^1.4.0",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@ianvs/prettier-plugin-sort-imports": "^4.2.1",
-    "@types/lodash": "^4.14.202",
-    "@types/node": "20.11.20",
+    "@types/lodash": "^4.17.4",
+    "@types/node": "20.12.12",
     "@types/nprogress": "^0.2.3",
-    "@types/react": "18.2.60",
-    "@types/react-dom": "18.2.19",
+    "@types/react": "18.3.2",
+    "@types/react-dom": "18.3.0",
     "@types/react-modal": "^3.16.3",
     "@types/react-scroll": "^1.8.10",
     "@types/styled-components": "^5.1.34",
@@ -60,12 +74,20 @@
     "@types/styled-system__css": "^5.0.21",
     "@types/styled-system__theme-get": "^5.0.4",
     "autoprefixer": "^10.4.19",
+    "drizzle-kit": "^0.21.2",
+    "eslint": "^9.2.0",
+    "eslint-config-next": "^14.2.3",
+    "eslint-plugin-drizzle": "^0.2.3",
+    "pg": "^8.11.5",
     "postcss": "^8.4.38",
     "prettier": "^3.2.5",
     "prettier-plugin-packagejson": "^2.5.0",
     "prettier-plugin-tailwindcss": "^0.5.14",
+    "supabase": "^1.167.4",
     "tailwindcss": "^3.4.3",
-    "typescript": "5.3.3"
+    "typescript": "5.4.5"
   },
-  "license": "https://themeforest.net/licenses/standard"
+  "ct3aMetadata": {
+    "initVersion": "7.33.1"
+  }
 }
